galaxys2-common: disable dexpreopt
Dexpreopt seems to introduce problems with wifi which causes soft-bootloops to occur.
Disabling dexpreopt explicitly for now, as we are having such soft-bootloop problem.
As a side note, there's no such problem on lineage-16.0 and lineage-17.1, so it is safe to remain enabled in newer android builds. lineage-15.1 is however, untested.

Change-Id: I541c36edd1ed345a9b17d3760dda0598ba363e6c