galaxys2-common: reconfigure wifi configuration (2)
Wifi is working now, however is not exactly because it is using fallback firmware path from the kernel, as the previously set firmware path is incorrect.
Point the firmware path for both STA and AP to the correct location, and also add a configuration line as per changes in libhardware_legacy.

Change-Id: Id62150045d66c889990d01b34f5920c6650a9539