Revert "galaxys2-common: replace cbd with replicant ipc-modem"
According to Replicant's website, SIM PIN is to be passed over to ipc-modem using command line parameter. This should explain the behaviour of RIL not working after enabling SIM PIN.
Reverting back to proprietary implementation of cbd. This left us with just the libsecril-client being the only open-source implemenation we are using.

This reverts commit 5355fbcc15350f7d9763449da170b7bb118f7a15.

Change-Id: I0ed5261164978f21817517fb498fd1222e90a585

Conflicts:
	common.mk
	configs/selinux/ipc-modem.te