galaxys2-common: disable anti-falsing classifier on the lockscreen
This causes user a hard time to unlock the lockscreen because our old touchscreen is considered noisy.
Compare with a new device with pointer location enabled, and notice the difference in the smoothness of line.
The more the edges are, the higher possibility the touch is rejected for unlocking.

Refer: https://review.lineageos.org/c/LineageOS/android_device_samsung_msm8974-common/+/273665