[HACK] galaxys2-common: build telephony-ext
telephony-ext is a qualcomm specific, however is being introduced in frameworks/opt/telephony globally.
According to hardware/custom/telephony, telephony-ext if is being included in build, it must be added to PRODUCT_BOOT_JARS.
In our case, as our target device does not include telephony-ext in device tree configuration, it is missing from PRODUCT_BOOT_JARS and thus causing dexpreopt to fail.

As a hack, we build telephony-ext and add it to PRODUCT_BOOT_JARS regardless of whether we are using qualcomm RIL or not. This avoids the need of modifying sources.

Refer:
1. https://review.potatoproject.co/plugins/gitiles/PotatoProject/hardware_custom_telephony/+/refs/heads/baked-release
2. https://github.com/PotatoProject/frameworks_opt_telephony/commit/f43623a623c2c41e18fdedeff5cdcf2d6c3def97
3. http://gerrit.aospextended.com/plugins/gitiles/AospExtended/platform_vendor_aosp/+/refs/heads/9.x/config/telephony.mk
4. https://hastebin.com/opinedemus.pl